refactor(app): remove duplicate handlers
