+++ apps/campfire/src/useDirectiveHandlers.ts
@@ -1,17 +1,290 @@
 import { useMemo } from 'react'
-import { handleDirective } from './directives/handleDirective'
+import { SKIP } from 'unist-util-visit'
+import { compile } from 'expression-eval'
+import { toString } from 'mdast-util-to-string'
+import type { Text, Parent } from 'mdast'
+import type { ContainerDirective } from 'mdast-util-directive'
+import { useGameStore } from '@/packages/use-game-store'
+import {
+  resolveIf,
+  evalCondition,
+  isRange,
+  clamp,
+  parseNumericValue,
+  parseRange,
+  getRandomItem,
+  getRandomInt,
+  ensureKey,
+  removeNode,
+  type DirectiveNode,
+  type RangeValue
+} from './directives/helpers'
+import type {
+  DirectiveHandler,
+  DirectiveHandlerResult
+} from '@/packages/remark-campfire'
 
-export const useDirectiveHandlers = () =>
-  useMemo(
+export const useDirectiveHandlers = () => {
+  const handleSet = (
+    directive: DirectiveNode,
+    parent: Parent | undefined,
+    index: number | undefined,
+    lock = false
+  ): DirectiveHandlerResult => {
+    const typeParam = (toString(directive).trim() || 'string').toLowerCase()
+    const attrs = directive.attributes
+    const safe: Record<string, unknown> = {}
+
+    const parseValue = (value: string): unknown => {
+      switch (typeParam) {
+        case 'number': {
+          let evaluated: unknown = value
+          try {
+            const fn = compile(value)
+            evaluated = fn(useGameStore.getState().gameData)
+          } catch {
+            // fall back to raw value when evaluation fails
+          }
+          if (typeof evaluated === 'number') return evaluated
+          const num = parseFloat(String(evaluated))
+          return Number.isNaN(num) ? 0 : num
+        }
+        case 'boolean':
+          return value === 'true'
+        case 'object':
+          try {
+            return JSON.parse(value)
+          } catch {
+            return {}
+          }
+        case 'range':
+          try {
+            const parsed = JSON.parse(value)
+            if (parsed && typeof parsed === 'object') {
+              return parseRange(parsed)
+            }
+            const num =
+              typeof parsed === 'number' ? parsed : parseFloat(String(parsed))
+            return Number.isNaN(num) ? 0 : num
+          } catch {
+            const num = parseFloat(value)
+            return Number.isNaN(num) ? 0 : num
+          }
+        case 'string':
+        default:
+          return value
+      }
+    }
+
+    if (attrs && typeof attrs === 'object') {
+      for (const [key, value] of Object.entries(attrs)) {
+        if (typeof value === 'string') {
+          const parsed = parseValue(value)
+          const current = useGameStore.getState().gameData[key]
+          if (isRange(current)) {
+            if (typeof parsed === 'number') {
+              safe[key] = {
+                ...current,
+                value: clamp(parsed, current.lower, current.upper)
+              }
+            } else if (isRange(parsed)) {
+              safe[key] = {
+                lower: parsed.lower,
+                upper: parsed.upper,
+                value: clamp(parsed.value, parsed.lower, parsed.upper)
+              }
+            } else {
+              safe[key] = parsed
+            }
+          } else {
+            safe[key] = parsed
+          }
+        }
+      }
+    }
+
+    if (Object.keys(safe).length > 0) {
+      const state = useGameStore.getState()
+      state.setGameData(safe)
+      if (lock) {
+        for (const k of Object.keys(safe)) {
+          state.lockKey(k)
+        }
+      }
+    }
+
+    if (parent && typeof index === 'number') {
+      parent.children.splice(index, 1)
+      return index
+    }
+  }
+
+  const handleGet: DirectiveHandler = (directive, parent, index) => {
+    const expr: string =
+      toString(directive) || Object.keys(directive.attributes || {})[0] || ''
+    let value: unknown
+    try {
+      const fn = compile(expr)
+      value = fn(useGameStore.getState().gameData)
+    } catch {
+      value = (useGameStore.getState().gameData as any)[expr]
+    }
+    if (isRange(value)) {
+      value = value.value
+    }
+    const textNode: Text = {
+      type: 'text',
+      value: value == null ? '' : String(value)
+    }
+    if (parent && typeof index === 'number') {
+      parent.children.splice(index, 1, textNode)
+      return index
+    }
+  }
+
+  const handleRandom: DirectiveHandler = (directive, parent, index) => {
+    const attrs = directive.attributes || {}
+    const key = ensureKey((attrs as Record<string, unknown>).key, parent, index)
+    if (!key) return index
+
+    let value: unknown
+
+    const optionsAttr =
+      (attrs as Record<string, unknown>).options ??
+      (attrs as Record<string, unknown>).from
+    if (typeof optionsAttr === 'string') {
+      const options = optionsAttr
+        .split(',')
+        .map(s => s.trim())
+        .filter(Boolean)
+      value = getRandomItem(options)
+    } else {
+      const minRaw = (attrs as Record<string, unknown>).min
+      const maxRaw = (attrs as Record<string, unknown>).max
+      const min =
+        typeof minRaw === 'number'
+          ? minRaw
+          : minRaw == null
+            ? undefined
+            : parseFloat(String(minRaw))
+      const max =
+        typeof maxRaw === 'number'
+          ? maxRaw
+          : maxRaw == null
+            ? undefined
+            : parseFloat(String(maxRaw))
+      if (
+        typeof min === 'number' &&
+        !Number.isNaN(min) &&
+        typeof max === 'number' &&
+        !Number.isNaN(max)
+      ) {
+        value = getRandomInt(min, max)
+      }
+    }
+
+    if (value !== undefined) {
+      useGameStore.getState().setGameData({ [key]: value })
+    }
+
+    const removed = removeNode(parent, index)
+    if (typeof removed === 'number') return removed
+  }
+
+  const handleIncrement = (
+    directive: DirectiveNode,
+    parent: Parent | undefined,
+    index: number | undefined,
+    sign = 1
+  ): DirectiveHandlerResult => {
+    const attrs = directive.attributes || {}
+    const key = ensureKey((attrs as Record<string, unknown>).key, parent, index)
+    if (!key) return index
+    const amountRaw = (attrs as Record<string, unknown>).amount
+
+    let amount: number = 1
+    if (typeof amountRaw === 'number') {
+      amount = amountRaw
+    } else if (typeof amountRaw === 'string') {
+      let evaluated: unknown = amountRaw
+      try {
+        const fn = compile(amountRaw)
+        evaluated = fn(useGameStore.getState().gameData)
+      } catch {
+        // ignore
+      }
+      amount =
+        typeof evaluated === 'number'
+          ? evaluated
+          : parseNumericValue(evaluated, 1)
+    }
+
+    amount *= sign
+
+    const state = useGameStore.getState()
+    const current = state.gameData[key]
+    if (isRange(current)) {
+      state.setGameData({
+        [key]: {
+          ...current,
+          value: clamp(current.value + amount, current.lower, current.upper)
+        }
+      })
+    } else {
+      const base = parseNumericValue(current, 0)
+      state.setGameData({ [key]: base + amount })
+    }
+
+    const removed = removeNode(parent, index)
+    if (typeof removed === 'number') return removed
+  }
+
+  const handleUnset: DirectiveHandler = (directive, parent, index) => {
+    const attrs = directive.attributes || {}
+    const key = ensureKey(
+      (attrs as Record<string, unknown>).key ?? toString(directive),
+      parent,
+      index
+    )
+    if (!key) return index
+
+    useGameStore.getState().unsetGameData(key)
+
+    return removeNode(parent, index)
+  }
+
+  const handleIf: DirectiveHandler = (directive, parent, index) => {
+    if (!parent || typeof index !== 'number') return
+    const ifDirective = directive as ContainerDirective
+    const replacement = resolveIf(ifDirective, evalCondition)
+    parent.children.splice(index, 1, ...replacement)
+    return [SKIP, index]
+  }
+
+  return useMemo(
     () => ({
-      set: handleDirective,
-      setOnce: handleDirective,
-      get: handleDirective,
-      random: handleDirective,
-      increment: handleDirective,
-      decrement: handleDirective,
-      unset: handleDirective,
-      if: handleDirective
+      set: (d: DirectiveNode, p: Parent | undefined, i: number | undefined) =>
+        handleSet(d, p, i, false),
+      setOnce: (
+        d: DirectiveNode,
+        p: Parent | undefined,
+        i: number | undefined
+      ) => handleSet(d, p, i, true),
+      get: handleGet,
+      random: handleRandom,
+      increment: (
+        d: DirectiveNode,
+        p: Parent | undefined,
+        i: number | undefined
+      ) => handleIncrement(d, p, i, 1),
+      decrement: (
+        d: DirectiveNode,
+        p: Parent | undefined,
+        i: number | undefined
+      ) => handleIncrement(d, p, i, -1),
+      unset: handleUnset,
+      if: handleIf
     }),
     []
   )
+}
